Refactoring the event editor form to a reusable component
This will make the Edit Event screen easier
