+++ src/AzureOpenAIProxy.Management/Components/Pages/EventAdd.razor
@@ -1,47 +1,11 @@
 @page "/events/new"
 @using System.ComponentModel.DataAnnotations
+@using AzureOpenAIProxy.Management.Components.EventManagement
+
 @namespace AzureOpenAIProxy.Management.Components.Pages
 
 <MudStack>
     <h1>Create a new event</h1>
 
-    <EditForm OnValidSubmit="@HandleValidSubmit" EditContext="@editContext" @bind-Disabled="isSubmitting">
-        <ValidationSummary />
-        <DataAnnotationsValidator />
-
-        <MudCard>
-            <MudCardContent>
-                <MudTextField Label="Event Name" @bind-Value="Model.Name" For="@(() => Model.Name)" />
-                <MudTextField Label="Event URL text" @bind-Value="Model.UrlText" For="@(() => Model.UrlText)" />
-                <MudTextField Label="Event URL" @bind-Value="Model.Url" InputType="InputType.Url"
-                    For="@(() => Model.Url)" />
-                <MudTextField Label="Organiser Name" @bind-Value="Model.OrganizerName"
-                    For="@(() => Model.OrganizerName)" />
-                <MudTextField Label="Organiser Email" @bind-Value="Model.OrganizerEmail" InputType="InputType.Email"
-                    For="@(() => Model.OrganizerEmail)" />
-                <div class="d-flex">
-                    <MudTextField Label="Start Date" @bind-Value="Model.Start" InputType="InputType.DateTimeLocal"
-                        For="@(() => Model.Start)" Format="s" />
-                    <MudTextField Label="End Date" @bind-Value="Model.End" InputType="InputType.DateTimeLocal"
-                        For="@(() => Model.End)" Format="s" />
-                </div>
-                <MudTextField Label="Description" @bind-Value="Model.Description" Lines="10"
-                    For="@(() => Model.Description)" />
-                <div class="d-flex">
-                    <MudNumericField Label="Max Tokens" @bind-Value="Model.MaxTokenCap" Min="0"
-                        For="@(() => Model.MaxTokenCap)" />
-                    <MudNumericField Label="Daily Request Cap" @bind-Value="Model.DailyRequestCap" Min="0"
-                        For="@(() => Model.DailyRequestCap)" />
-                </div>
-                <div class="d-flex">
-                    <MudSwitch @bind-Value="Model.Active" Label="Active" Color="Color.Primary" />
-                    <MudSwitch @bind-Value="Model.SingleCode" Label="Single Code" />
-                </div>
-            </MudCardContent>
-            <MudCardActions>
-                <MudButton ButtonType="ButtonType.Submit" Variant="Variant.Filled" Color="@(isSubmitting ? Color.Info : Color.Primary)" Class="ml-auto"
-                    StartIcon="@Icons.Material.Filled.Save" Disabled="@isSubmitting">@(isSubmitting ? "Saving" : "Save")</MudButton>
-            </MudCardActions>
-        </MudCard>
-    </EditForm>
+    <EventEditor OnValidSubmit="@HandleValidSubmit" />
 </MudStack>
